Fold ListWidget<T> into ListWidget.cs
Both records share the same file now — the legacy non-generic record is
declared first (with its [Obsolete] attribute) and the generic ListWidget<T>
follows.
